refactor(reference-service)!: rename application layer to service, merge order use cases
Rename src/reference_service/application/ to service/ across the code,
import-linter contracts, mypy strict overrides, tests and README. The
application-independence import-linter contract is renamed to
service-independence to match what it guards.

Consolidate place_order.py, get_order.py and dtos.py into a single
service/order.py (PlaceOrderLine, PlaceOrderCommand, PlaceOrder,
GetOrder) since all four belong to the same Order aggregate. Merge the
corresponding unit tests into tests/unit/test_order_service.py for the
same reason: one file per aggregate, on both sides.

No behaviour change: signatures, validators (including the
currency-consistency check) and test assertions are unchanged.
